Module aws_sdk_workspaces::client::fluent_builders

Utilities to ergonomically construct a request to the service.

Fluent builders are created through the Client by calling one if its operation methods. After parameters are set using the builder methods, the send method can be called to initiate the request.

Structs

AssociateConnectionAlias

Fluent builder constructing a request to AssociateConnectionAlias.

AssociateIpGroups

Fluent builder constructing a request to AssociateIpGroups.

AuthorizeIpRules

Fluent builder constructing a request to AuthorizeIpRules.

CopyWorkspaceImage

Fluent builder constructing a request to CopyWorkspaceImage.

CreateConnectionAlias

Fluent builder constructing a request to CreateConnectionAlias.

CreateIpGroup

Fluent builder constructing a request to CreateIpGroup.

CreateTags

Fluent builder constructing a request to CreateTags.

CreateUpdatedWorkspaceImage

Fluent builder constructing a request to CreateUpdatedWorkspaceImage.

CreateWorkspaceBundle

Fluent builder constructing a request to CreateWorkspaceBundle.

CreateWorkspaces

Fluent builder constructing a request to CreateWorkspaces.

DeleteConnectionAlias

Fluent builder constructing a request to DeleteConnectionAlias.

DeleteIpGroup

Fluent builder constructing a request to DeleteIpGroup.

DeleteTags

Fluent builder constructing a request to DeleteTags.

DeleteWorkspaceBundle

Fluent builder constructing a request to DeleteWorkspaceBundle.

DeleteWorkspaceImage

Fluent builder constructing a request to DeleteWorkspaceImage.

DeregisterWorkspaceDirectory

Fluent builder constructing a request to DeregisterWorkspaceDirectory.

DescribeAccount

Fluent builder constructing a request to DescribeAccount.

DescribeAccountModifications

Fluent builder constructing a request to DescribeAccountModifications.

DescribeClientProperties

Fluent builder constructing a request to DescribeClientProperties.

DescribeConnectionAliasPermissions

Fluent builder constructing a request to DescribeConnectionAliasPermissions.

DescribeConnectionAliases

Fluent builder constructing a request to DescribeConnectionAliases.

DescribeIpGroups

Fluent builder constructing a request to DescribeIpGroups.

DescribeTags

Fluent builder constructing a request to DescribeTags.

DescribeWorkspaceBundles

Fluent builder constructing a request to DescribeWorkspaceBundles.

DescribeWorkspaceDirectories

Fluent builder constructing a request to DescribeWorkspaceDirectories.

DescribeWorkspaceImagePermissions

Fluent builder constructing a request to DescribeWorkspaceImagePermissions.

DescribeWorkspaceImages

Fluent builder constructing a request to DescribeWorkspaceImages.

DescribeWorkspaceSnapshots

Fluent builder constructing a request to DescribeWorkspaceSnapshots.

DescribeWorkspaces

Fluent builder constructing a request to DescribeWorkspaces.

DescribeWorkspacesConnectionStatus

Fluent builder constructing a request to DescribeWorkspacesConnectionStatus.

DisassociateConnectionAlias

Fluent builder constructing a request to DisassociateConnectionAlias.

DisassociateIpGroups

Fluent builder constructing a request to DisassociateIpGroups.

ImportWorkspaceImage

Fluent builder constructing a request to ImportWorkspaceImage.

ListAvailableManagementCidrRanges

Fluent builder constructing a request to ListAvailableManagementCidrRanges.

MigrateWorkspace

Fluent builder constructing a request to MigrateWorkspace.

ModifyAccount

Fluent builder constructing a request to ModifyAccount.

ModifyClientProperties

Fluent builder constructing a request to ModifyClientProperties.

ModifySelfservicePermissions

Fluent builder constructing a request to ModifySelfservicePermissions.

ModifyWorkspaceAccessProperties

Fluent builder constructing a request to ModifyWorkspaceAccessProperties.

ModifyWorkspaceCreationProperties

Fluent builder constructing a request to ModifyWorkspaceCreationProperties.

ModifyWorkspaceProperties

Fluent builder constructing a request to ModifyWorkspaceProperties.

ModifyWorkspaceState

Fluent builder constructing a request to ModifyWorkspaceState.

RebootWorkspaces

Fluent builder constructing a request to RebootWorkspaces.

RebuildWorkspaces

Fluent builder constructing a request to RebuildWorkspaces.

RegisterWorkspaceDirectory

Fluent builder constructing a request to RegisterWorkspaceDirectory.

RestoreWorkspace

Fluent builder constructing a request to RestoreWorkspace.

RevokeIpRules

Fluent builder constructing a request to RevokeIpRules.

StartWorkspaces

Fluent builder constructing a request to StartWorkspaces.

StopWorkspaces

Fluent builder constructing a request to StopWorkspaces.

TerminateWorkspaces

Fluent builder constructing a request to TerminateWorkspaces.

UpdateConnectionAliasPermission

Fluent builder constructing a request to UpdateConnectionAliasPermission.

UpdateRulesOfIpGroup

Fluent builder constructing a request to UpdateRulesOfIpGroup.

UpdateWorkspaceBundle

Fluent builder constructing a request to UpdateWorkspaceBundle.

UpdateWorkspaceImagePermission

Fluent builder constructing a request to UpdateWorkspaceImagePermission.