Enum aws_sdk_workspaces::Error

#[non_exhaustive]
pub enum Error {

    AccessDeniedException(AccessDeniedException),
    InvalidParameterValuesException(InvalidParameterValuesException),
    InvalidResourceStateException(InvalidResourceStateException),
    OperationInProgressException(OperationInProgressException),
    OperationNotSupportedException(OperationNotSupportedException),
    ResourceAlreadyExistsException(ResourceAlreadyExistsException),
    ResourceAssociatedException(ResourceAssociatedException),
    ResourceCreationFailedException(ResourceCreationFailedException),
    ResourceLimitExceededException(ResourceLimitExceededException),
    ResourceNotFoundException(ResourceNotFoundException),
    ResourceUnavailableException(ResourceUnavailableException),
    UnsupportedNetworkConfigurationException(UnsupportedNetworkConfigurationException),
    UnsupportedWorkspaceConfigurationException(UnsupportedWorkspaceConfigurationException),
    WorkspacesDefaultRoleNotFoundException(WorkspacesDefaultRoleNotFoundException),
    Unhandled(Box<dyn Error + Send + Sync + 'static>),

}

All possible error types for this service.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

AccessDeniedException(AccessDeniedException)

The user is not authorized to access a resource.

Tuple Fields of AccessDeniedException

0: AccessDeniedException

InvalidParameterValuesException(InvalidParameterValuesException)

One or more parameter values are not valid.

Tuple Fields of InvalidParameterValuesException

0: InvalidParameterValuesException

InvalidResourceStateException(InvalidResourceStateException)

The state of the resource is not valid for this operation.

Tuple Fields of InvalidResourceStateException

0: InvalidResourceStateException

OperationInProgressException(OperationInProgressException)

The properties of this WorkSpace are currently being modified. Try again in a moment.

Tuple Fields of OperationInProgressException

0: OperationInProgressException

OperationNotSupportedException(OperationNotSupportedException)

This operation is not supported.

Tuple Fields of OperationNotSupportedException

0: OperationNotSupportedException

ResourceAlreadyExistsException(ResourceAlreadyExistsException)

The specified resource already exists.

Tuple Fields of ResourceAlreadyExistsException

0: ResourceAlreadyExistsException

ResourceAssociatedException(ResourceAssociatedException)

The resource is associated with a directory.

Tuple Fields of ResourceAssociatedException

0: ResourceAssociatedException

ResourceCreationFailedException(ResourceCreationFailedException)

The resource could not be created.

Tuple Fields of ResourceCreationFailedException

0: ResourceCreationFailedException

ResourceLimitExceededException(ResourceLimitExceededException)

Your resource limits have been exceeded.

Tuple Fields of ResourceLimitExceededException

0: ResourceLimitExceededException

ResourceNotFoundException(ResourceNotFoundException)

The resource could not be found.

Tuple Fields of ResourceNotFoundException

0: ResourceNotFoundException

ResourceUnavailableException(ResourceUnavailableException)

The specified resource is not available.

Tuple Fields of ResourceUnavailableException

0: ResourceUnavailableException

UnsupportedNetworkConfigurationException(UnsupportedNetworkConfigurationException)

The configuration of this network is not supported for this operation, or your network configuration conflicts with the Amazon WorkSpaces management network IP range. For more information, see Configure a VPC for Amazon WorkSpaces.

Tuple Fields of UnsupportedNetworkConfigurationException

0: UnsupportedNetworkConfigurationException

UnsupportedWorkspaceConfigurationException(UnsupportedWorkspaceConfigurationException)

The configuration of this WorkSpace is not supported for this operation. For more information, see Required Configuration and Service Components for WorkSpaces .

Tuple Fields of UnsupportedWorkspaceConfigurationException

0: UnsupportedWorkspaceConfigurationException

WorkspacesDefaultRoleNotFoundException(WorkspacesDefaultRoleNotFoundException)

The workspaces_DefaultRole role could not be found. If this is the first time you are registering a directory, you will need to create the workspaces_DefaultRole role before you can register a directory. For more information, see Creating the workspaces_DefaultRole Role.

Tuple Fields of WorkspacesDefaultRoleNotFoundException

0: WorkspacesDefaultRoleNotFoundException

Unhandled(Box<dyn Error + Send + Sync + 'static>)

An unhandled error occurred.

Tuple Fields of Unhandled

0: Box<dyn Error + Send + Sync + 'static>