Module aws_sdk_workspaces::error

Modules

access_denied_exception

See AccessDeniedException

invalid_parameter_values_exception

See InvalidParameterValuesException

invalid_resource_state_exception

See InvalidResourceStateException

operation_in_progress_exception

See OperationInProgressException

operation_not_supported_exception

See OperationNotSupportedException

resource_already_exists_exception

See ResourceAlreadyExistsException

resource_associated_exception

See ResourceAssociatedException

resource_creation_failed_exception

See ResourceCreationFailedException

resource_limit_exceeded_exception

See ResourceLimitExceededException

resource_not_found_exception

See ResourceNotFoundException

resource_unavailable_exception

See ResourceUnavailableException

unsupported_network_configuration_exception

See UnsupportedNetworkConfigurationException

unsupported_workspace_configuration_exception

See UnsupportedWorkspaceConfigurationException

workspaces_default_role_not_found_exception

See WorkspacesDefaultRoleNotFoundException

Structs

AccessDeniedException

The user is not authorized to access a resource.

AssociateConnectionAliasError

AssociateIpGroupsError

AuthorizeIpRulesError

CopyWorkspaceImageError

CreateConnectionAliasError

CreateIpGroupError

CreateTagsError

CreateWorkspaceBundleError

CreateWorkspacesError

DeleteConnectionAliasError

DeleteIpGroupError

DeleteTagsError

DeleteWorkspaceBundleError

DeleteWorkspaceImageError

DeregisterWorkspaceDirectoryError

DescribeAccountError

DescribeAccountModificationsError

DescribeClientPropertiesError

DescribeConnectionAliasPermissionsError

DescribeConnectionAliasesError

DescribeIpGroupsError

DescribeTagsError

DescribeWorkspaceBundlesError

DescribeWorkspaceDirectoriesError

DescribeWorkspaceImagePermissionsError

DescribeWorkspaceImagesError

DescribeWorkspaceSnapshotsError

DescribeWorkspacesConnectionStatusError

DescribeWorkspacesError

DisassociateConnectionAliasError

DisassociateIpGroupsError

ImportWorkspaceImageError

InvalidParameterValuesException

One or more parameter values are not valid.

InvalidResourceStateException

The state of the resource is not valid for this operation.

ListAvailableManagementCidrRangesError

MigrateWorkspaceError

ModifyAccountError

ModifyClientPropertiesError

ModifySelfservicePermissionsError

ModifyWorkspaceAccessPropertiesError

ModifyWorkspaceCreationPropertiesError

ModifyWorkspacePropertiesError

ModifyWorkspaceStateError

OperationInProgressException

The properties of this WorkSpace are currently being modified. Try again in a moment.

OperationNotSupportedException

This operation is not supported.

RebootWorkspacesError

RebuildWorkspacesError

RegisterWorkspaceDirectoryError

ResourceAlreadyExistsException

The specified resource already exists.

ResourceAssociatedException

The resource is associated with a directory.

ResourceCreationFailedException

The resource could not be created.

ResourceLimitExceededException

Your resource limits have been exceeded.

ResourceNotFoundException

The resource could not be found.

ResourceUnavailableException

The specified resource is not available.

RestoreWorkspaceError

RevokeIpRulesError

StartWorkspacesError

StopWorkspacesError

TerminateWorkspacesError

UnsupportedNetworkConfigurationException

The configuration of this network is not supported for this operation, or your network configuration

UnsupportedWorkspaceConfigurationException

The configuration of this WorkSpace is not supported for this operation. For more information, see

UpdateConnectionAliasPermissionError

UpdateRulesOfIpGroupError

UpdateWorkspaceBundleError

UpdateWorkspaceImagePermissionError

WorkspacesDefaultRoleNotFoundException

The workspaces_DefaultRole role could not be found. If this is the first time you are registering a directory, you

Enums

AssociateConnectionAliasErrorKind

AssociateIpGroupsErrorKind

AuthorizeIpRulesErrorKind

CopyWorkspaceImageErrorKind

CreateConnectionAliasErrorKind

CreateIpGroupErrorKind

CreateTagsErrorKind

CreateWorkspaceBundleErrorKind

CreateWorkspacesErrorKind

DeleteConnectionAliasErrorKind

DeleteIpGroupErrorKind

DeleteTagsErrorKind

DeleteWorkspaceBundleErrorKind

DeleteWorkspaceImageErrorKind

DeregisterWorkspaceDirectoryErrorKind

DescribeAccountErrorKind

DescribeAccountModificationsErrorKind

DescribeClientPropertiesErrorKind

DescribeConnectionAliasPermissionsErrorKind

DescribeConnectionAliasesErrorKind

DescribeIpGroupsErrorKind

DescribeTagsErrorKind

DescribeWorkspaceBundlesErrorKind

DescribeWorkspaceDirectoriesErrorKind

DescribeWorkspaceImagePermissionsErrorKind

DescribeWorkspaceImagesErrorKind

DescribeWorkspaceSnapshotsErrorKind

DescribeWorkspacesConnectionStatusErrorKind

DescribeWorkspacesErrorKind

DisassociateConnectionAliasErrorKind

DisassociateIpGroupsErrorKind

ImportWorkspaceImageErrorKind

ListAvailableManagementCidrRangesErrorKind

MigrateWorkspaceErrorKind

ModifyAccountErrorKind

ModifyClientPropertiesErrorKind

ModifySelfservicePermissionsErrorKind

ModifyWorkspaceAccessPropertiesErrorKind

ModifyWorkspaceCreationPropertiesErrorKind

ModifyWorkspacePropertiesErrorKind

ModifyWorkspaceStateErrorKind

RebootWorkspacesErrorKind

RebuildWorkspacesErrorKind

RegisterWorkspaceDirectoryErrorKind

RestoreWorkspaceErrorKind

RevokeIpRulesErrorKind

StartWorkspacesErrorKind

StopWorkspacesErrorKind

TerminateWorkspacesErrorKind

UpdateConnectionAliasPermissionErrorKind

UpdateRulesOfIpGroupErrorKind

UpdateWorkspaceBundleErrorKind

UpdateWorkspaceImagePermissionErrorKind