Module aws_sdk_workspaces::model

Modules

account_modification

See AccountModification

client_properties

See ClientProperties

client_properties_result

See ClientPropertiesResult

compute_type

See ComputeType

connection_alias

See ConnectionAlias

connection_alias_association

See ConnectionAliasAssociation

connection_alias_permission

See ConnectionAliasPermission

default_workspace_creation_properties

See DefaultWorkspaceCreationProperties

failed_create_workspace_request

See FailedCreateWorkspaceRequest

failed_workspace_change_request

See FailedWorkspaceChangeRequest

image_permission

See ImagePermission

ip_rule_item

See IpRuleItem

modification_state

See ModificationState

operating_system

See OperatingSystem

reboot_request

See RebootRequest

rebuild_request

See RebuildRequest

root_storage

See RootStorage

selfservice_permissions

See SelfservicePermissions

snapshot

See Snapshot

start_request

See StartRequest

stop_request

See StopRequest

tag

See Tag

terminate_request

See TerminateRequest

user_storage

See UserStorage

workspace

See Workspace

workspace_access_properties

See WorkspaceAccessProperties

workspace_bundle

See WorkspaceBundle

workspace_connection_status

See WorkspaceConnectionStatus

workspace_creation_properties

See WorkspaceCreationProperties

workspace_directory

See WorkspaceDirectory

workspace_image

See WorkspaceImage

workspace_properties

See WorkspaceProperties

workspace_request

See WorkspaceRequest

workspaces_ip_group

See WorkspacesIpGroup

Structs

AccountModification

Describes a modification to the configuration of Bring Your Own License (BYOL) for the

ClientProperties

Describes an Amazon WorkSpaces client.

ClientPropertiesResult

Information about the Amazon WorkSpaces client.

ComputeType

Describes the compute type of the bundle.

ConnectionAlias

Describes a connection alias. Connection aliases are used for cross-Region redirection. For more information,

ConnectionAliasAssociation

Describes a connection alias association that is used for cross-Region redirection. For more information, see

ConnectionAliasPermission

Describes the permissions for a connection alias. Connection aliases are used for cross-Region redirection.

DefaultWorkspaceCreationProperties

Describes the default values that are used to create WorkSpaces. For more information, see Update Directory Details for Your WorkSpaces.

FailedCreateWorkspaceRequest

Describes a WorkSpace that cannot be created.

FailedWorkspaceChangeRequest

Describes a WorkSpace that could not be rebooted.

ImagePermission

Describes the AWS accounts that have been granted permission to use a shared image. For more

IpRuleItem

Describes a rule for an IP access control group.

ModificationState

Describes a WorkSpace modification.

OperatingSystem

The operating system that the image is running.

RebootRequest

Describes the information used to reboot a WorkSpace.

RebuildRequest

Describes the information used to rebuild a WorkSpace.

RootStorage

Describes the root volume for a WorkSpace bundle.

SelfservicePermissions

Describes the self-service permissions for a directory. For more information, see Enable Self-Service WorkSpace Management Capabilities for Your Users.

Snapshot

Describes a snapshot.

StartRequest

Information used to start a WorkSpace.

StopRequest

Describes the information used to stop a WorkSpace.

Tag

Describes a tag.

TerminateRequest

Describes the information used to terminate a WorkSpace.

UserStorage

Describes the user volume for a WorkSpace bundle.

Workspace

Describes a WorkSpace.

WorkspaceAccessProperties

The device types and operating systems that can be used to access a WorkSpace. For more

WorkspaceBundle

Describes a WorkSpace bundle.

WorkspaceConnectionStatus

Describes the connection status of a WorkSpace.

WorkspaceCreationProperties

Describes the default properties that are used for creating WorkSpaces. For more

WorkspaceDirectory

Describes a directory that is used with Amazon WorkSpaces.

WorkspaceImage

Describes a WorkSpace image.

WorkspaceProperties

Describes a WorkSpace.

WorkspaceRequest

Describes the information used to create a WorkSpace.

WorkspacesIpGroup

Describes an IP access control group.

Enums

AccessPropertyValue

Application

AssociationStatus

Compute

ConnectionAliasState

ConnectionState

NOTE: ConnectionState::Unknown has been renamed to ::UnknownValue.

DedicatedTenancyModificationStateEnum

DedicatedTenancySupportEnum

DedicatedTenancySupportResultEnum

ImageType

ModificationResourceEnum

ModificationStateEnum

OperatingSystemType

ReconnectEnum

RunningMode

TargetWorkspaceState

Tenancy

WorkspaceDirectoryState

WorkspaceDirectoryType

WorkspaceImageIngestionProcess

WorkspaceImageRequiredTenancy

WorkspaceImageState

WorkspaceState