Module aws_sdk_workspaces::operation

Structs

AssociateConnectionAlias

Associates the specified connection alias with the specified directory to enable cross-Region redirection.

AssociateIpGroups

Associates the specified IP access control group with the specified directory.

AuthorizeIpRules

Adds one or more rules to the specified IP access control group.

CopyWorkspaceImage

Copies the specified image from the specified Region to the current Region.

CreateConnectionAlias

Creates the specified connection alias for use with cross-Region redirection. For more information, see

CreateIpGroup

Creates an IP access control group.

CreateTags

Creates the specified tags for the specified WorkSpaces resource.

CreateWorkspaceBundle

Creates the specified WorkSpace bundle. For more information about creating WorkSpace bundles, see

CreateWorkspaces

Creates one or more WorkSpaces.

DeleteConnectionAlias

Deletes the specified connection alias. For more information, see

DeleteIpGroup

Deletes the specified IP access control group.

DeleteTags

Deletes the specified tags from the specified WorkSpaces resource.

DeleteWorkspaceBundle

Deletes the specified WorkSpace bundle. For more information about deleting WorkSpace bundles, see

DeleteWorkspaceImage

Deletes the specified image from your account. To delete an image, you must first delete

DeregisterWorkspaceDirectory

Deregisters the specified directory. This operation is asynchronous

DescribeAccount

Retrieves a list that describes the configuration of Bring Your Own License (BYOL) for

DescribeAccountModifications

Retrieves a list that describes modifications to the configuration of Bring Your Own

DescribeClientProperties

Retrieves a list that describes one or more specified Amazon WorkSpaces clients.

DescribeConnectionAliasPermissions

Describes the permissions that the owner of a connection alias has granted to another AWS account for

DescribeConnectionAliases

Retrieves a list that describes the connection aliases used for cross-Region redirection. For more information, see

DescribeIpGroups

Describes one or more of your IP access control groups.

DescribeTags

Describes the specified tags for the specified WorkSpaces resource.

DescribeWorkspaceBundles

Retrieves a list that describes the available WorkSpace bundles.

DescribeWorkspaceDirectories

Describes the available directories that are registered with

DescribeWorkspaceImagePermissions

Describes the permissions that the owner of an image has granted to other

DescribeWorkspaceImages

Retrieves a list that describes one or more specified images, if the image identifiers

DescribeWorkspaceSnapshots

Describes the snapshots for the specified WorkSpace.

DescribeWorkspaces

Describes the specified WorkSpaces.

DescribeWorkspacesConnectionStatus

Describes the connection status of the specified WorkSpaces.

DisassociateConnectionAlias

Disassociates a connection alias from a directory. Disassociating a connection alias disables cross-Region

DisassociateIpGroups

Disassociates the specified IP access control group from the specified directory.

ImportWorkspaceImage

Imports the specified Windows 10 Bring Your Own License (BYOL) image into Amazon

ListAvailableManagementCidrRanges

Retrieves a list of IP address ranges, specified as IPv4 CIDR blocks, that you can use

MigrateWorkspace

Migrates a WorkSpace from one operating system or bundle type to another, while retaining the data on the user volume.

ModifyAccount

Modifies the configuration of Bring Your Own License (BYOL) for the specified

ModifyClientProperties

Modifies the properties of the specified Amazon WorkSpaces clients.

ModifySelfservicePermissions

Modifies the self-service WorkSpace management capabilities for your users. For more

ModifyWorkspaceAccessProperties

Specifies which devices and operating systems users can use to access their WorkSpaces.

ModifyWorkspaceCreationProperties

Modify the default properties used to create WorkSpaces.

ModifyWorkspaceProperties

Modifies the specified WorkSpace properties. For important information about how

ModifyWorkspaceState

Sets the state of the specified WorkSpace.

RebootWorkspaces

Reboots the specified WorkSpaces.

RebuildWorkspaces

Rebuilds the specified WorkSpace.

RegisterWorkspaceDirectory

Registers the specified directory. This operation is asynchronous

RestoreWorkspace

Restores the specified WorkSpace to its last known healthy state.

RevokeIpRules

Removes one or more rules from the specified IP access control group.

StartWorkspaces

Starts the specified WorkSpaces.

StopWorkspaces

Stops the specified WorkSpaces.

TerminateWorkspaces

Terminates the specified WorkSpaces.

UpdateConnectionAliasPermission

Shares or unshares a connection alias with one account by specifying whether that account has permission to

UpdateRulesOfIpGroup

Replaces the current rules of the specified IP access control group with the specified

UpdateWorkspaceBundle

Updates a WorkSpace bundle with a new image. For more information about updating WorkSpace bundles, see

UpdateWorkspaceImagePermission

Shares or unshares an image with one account in the same AWS Region by specifying whether that account has